requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.0.0
aiosqlite==0.19.0
schedule==1.2.0
python-dotenv==1.0.0
